requirements.txt
aniso8601==4.1.0
awsebcli==3.14.10
blessed==1.15.0
botocore==1.12.86
cached-property==1.5.1
cement==2.8.2
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.3.9
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==1.0.2
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
MarkupSafe==1.1.0
mysql-connector==2.1.6
pathspec==0.5.5
psycopg2==2.7.7
PyMySQL==0.9.3
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
SQLAlchemy==1.2.17
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.1
wcwidth==0.1.7
websocket-client==0.54.0
Werkzeug==0.14.1
